// frontend/src/api/generated/models/healthMetric.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type HealthMetric = typeof HealthMetric[keyof typeof HealthMetric];


export const HealthMetric = {
  Hrv: 'Hrv',
  RestingHeartRate: 'RestingHeartRate',
  SleepDuration: 'SleepDuration',
  SleepScore: 'SleepScore',
  AvgSleepingHeartRate: 'AvgSleepingHeartRate',
  SpO2: 'SpO2',
} as const;