// frontend/src/api/generated/models/healthMetricStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type HealthMetricStatus = typeof HealthMetricStatus[keyof typeof HealthMetricStatus];


export const HealthMetricStatus = {
  InRange: 'InRange',
  AboveRange: 'AboveRange',
  BelowRange: 'BelowRange',
  NoRange: 'NoRange',
  NoData: 'NoData',
} as const;