// frontend/src/api/generated/models/readinessFactor.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */

export type ReadinessFactor = typeof ReadinessFactor[keyof typeof ReadinessFactor];


export const ReadinessFactor = {
  Hrv: 'Hrv',
  RestingHeartRate: 'RestingHeartRate',
  Sleep: 'Sleep',
  TrainingLoad: 'TrainingLoad',
  Subjective: 'Subjective',
} as const;