// frontend/src/api/generated/models/healthMetricDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TrainCoach API
 * OpenAPI spec version: v1
 */
import type { HealthMetric } from './healthMetric';
import type { HealthMetricStatus } from './healthMetricStatus';

export interface HealthMetricDto {
  metric?: HealthMetric;
  /** @nullable */
  value?: number | null;
  /** @nullable */
  rangeLow?: number | null;
  /** @nullable */
  rangeHigh?: number | null;
  status?: HealthMetricStatus;
  isConcerning?: boolean;
}